backend/requirements.txt
# Paper Research Platform - FastAPI + DSPy Backend
# Telkom University Paper Research API

# Web Framework
fastapi
uvicorn

# Data Validation
pydantic
pydantic-settings

# AI Framework
dspy-ai

# JSON Processing (faster than standard json)
orjson

# Environment Variables
python-dotenv
